// packages/opentok-client-sdk/enums/connection-status.ts
export enum ConnectionStatus {
  Connecting = "connecting",
  ReConnecting = "reconnecting",
  Disconnected = "disconnected",
  Connected = "connected",
  Idle = "idle",
  Failed = "failed",
}
export enum SignalPermission {
  Idle = "Idle",
  Granted = "Granted",
  Denied = "Denied",
  Ignored = "Ignored",
  // Responded:Responded
}